+++ src/lib/components/EditableText.svelte
@@ -4,7 +4,7 @@
 	import TextField from './TextField.svelte';
 	import Dots from './Dots.svelte';
 	import Note from './Note.svelte';
-	import { addError, handle } from '../../routes/errors.svelte';
+	import { addError, handle } from '../../routes/feedback.svelte';
 	import { type ErrorID } from '$lib/data/CRUD';
 	import { ConfirmLabel, DeleteLabel, EditLabel } from './Labels';