+++ src/lib/components/Transactions.svelte
@@ -1,23 +1,43 @@
 <script lang="ts">
-	import type { TransactionRow, VenueID } from '$data/types';
+	import { page } from '$app/stores';
+	import type { CurrencyRow, TransactionRow, VenueID } from '$data/types';
+	import { getDB } from '$lib/data/CRUD';
+	import { getAuth } from '../../routes/Auth.svelte';
+	import { handle } from '../../routes/feedback.svelte';
+	import Button from './Button.svelte';
 	import Feedback from './Feedback.svelte';
 	import ScholarLink from './ScholarLink.svelte';
+	import Status from './Status.svelte';
 	import Table from './Table.svelte';
-	import Tag from './Tag.svelte';
 	import Tokens from './Tokens.svelte';
 	import VenueLink from './VenueLink.svelte';
 
 	let {
 		transactions,
-		venues
+		venues,
+		currency
 	}: {
 		transactions: TransactionRow[];
 		venues: { id: VenueID; title: string }[];
+		currency: CurrencyRow | undefined;
 	} = $props();
+
+	// Get the current user
+	const db = getDB();
+	const auth = getAuth();
+
+	// Editable if the user is the scholar being viewed.
+	let userid = $derived(auth.getUserID());
+	let editable = $derived(
+		userid !== null && currency !== undefined && currency.minters.includes(userid)
+	);
 </script>
 
 {#snippet row(transaction: TransactionRow)}
 	<tr>
+		<td>
+			<Status good={transaction.status === 'approved'}>{transaction.status}</Status>
+		</td>
 		<td
 			>{#if transaction.tokens === null}unknown{:else}<Tokens
 					amount={transaction.tokens.length}
@@ -40,6 +60,12 @@
 				></VenueLink>{/if}</td
 		>
 		<td>{transaction.purpose}</td>
+		<td>
+			{#if transaction.status === 'proposed' && editable && userid}<Button
+					tip="Approve this proposed transaction"
+					action={() => handle(db.approveTransaction(userid, transaction.id))}>Approve</Button
+				>{/if}
+		</td>
 	</tr>
 {/snippet}
 
@@ -48,10 +74,12 @@
 {:else}
 	<Table>
 		<tr>
+			<th>Status</th>
 			<th>Tokens</th>
 			<th>From</th>
 			<th>To</th>
 			<th>Purpose</th>
+			<th></th>
 		</tr>
 		{#each transactions.sort((a, b) => new Date(b.created).getTime() - new Date(a.created).getTime()) as transaction}
 			{@render row(transaction)}