+++ src/routes/currency/[id]/transactions/+page.ts
@@ -4,12 +4,19 @@ import type { PageLoad } from './$types';
 export const load: PageLoad = async ({ parent, params }) => {
 	const { supabase } = await parent();
 
-	// Get the scholar's most recent transactions.
+	// Get the currency's most recent transactions.
 	const { data: transactions, error: transactionsError } = await supabase
 		.from('transactions')
 		.select()
-		.or(`from_venue.eq.${params.id},to_venue.eq.${params.id}`);
+		.eq('currency', params.id);
 	if (transactionsError) console.log(transactionsError);
+
+	// Is the current scholar a minter on the venue?
+	const { data: currency, error: currencyError } = await supabase
+		.from('currencies')
+		.select()
+		.eq('id', params.id);
+	if (currencyError) console.log(currencyError);
 
 	const { data: venues, error: venueError } =
 		transactions === null