+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 ### Added
 
 - Allow scholars to gift tokens.
+- Added pattern for explicit success feedback.
 
 ### Maintenance